Tomato_disease_prediction_app/requirements.txt
keras
numpy
Pillow
streamlit
tensorflow